requirements.txt
requests==2.22.*
pyyaml==5.3.*
google-cloud-bigquery==2.7.*
google-cloud-pubsub==2.3.*